frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
pillow==10.1.0
